# app/Resources/api_doc.json
{
    "swagger": "2.0",
    "info": {
        "title": "Page API documentation",
        "version": "1.0.0"
    },
    "host": "rochedor.fr",
    "basePath": "/api",
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/speaker": {
            "get": {
                "summary": "Get requested speakers'list ordered by position",
                "responses": {
                    "200": {
                        "description": "The requested speakers"
                    }
                }
            },
            "post": {
                "summary": "Add a new speaker",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "title": {
                                    "type": "json_array"
                                },
                                "description": {
                                    "type": "json_array"
                                },
                                "image": {
                                    "type": "string"
                                },
                                "position": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The created speaker"
                    }
                }
            }
        },
        "/speaker/{id}/{version}": {
            "get": {
                "summary": "Get a speaker",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The speaker id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "description": "The speaker version",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested speaker"
                    },
                    "404": {
                        "description": "Speaker not found"
                    }
                }
            }
        },
        "/speaker/id": {
            "delete": {
                "summary": "Delete requested speaker",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "speaker id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The speaker is deleted"
                    },
                    "404": {
                        "description": "Speaker not found"
                    }
                }
            }
        },
        "/speaker/{id}": {
            "put": {
                "summary": "Edit requester speaker",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "title": {
                                    "type": "json_array"
                                },
                                "description": {
                                    "type": "json_array"
                                },
                                "image": {
                                    "type": "string"
                                },
                                "position": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "speaker id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "404": {
                        "description": "Speaker not found"
                    }
                }
            }
        },
        "/speaker/{id}/versions": {
            "get": {
                "summary": "Return all log entries for the selected speaker",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "the speaker ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested logs"
                    },
                    "404": {
                        "description": "Speaker not found"
                    }
                }
            },
            "put": {
                "summary": "Revert a speaker into the selected version",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "the speaker id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "description": "The speaker version",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested speaker at the selected version"
                    }
                }
            }
        },
        "/speaker/{id}/position/{position}": {
            "put": {
                "summary": "Change the position of a speaker and return a list with the new scheduling",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "the speaker id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "position",
                        "in": "path",
                        "description": "The speaker desired position",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of speaker with the new scheduling"
                    }
                }
            }
        }
    },
...
}